stream-deck: Adjust texture uploading of Original (gen1)
We must send at most half of the BMP image in the first package,
then send the other half. If we send more than half of the image
in the first package - which should always happen with the current
code - Stream Deck Original (gen1) ignored the extra stuff we sent,
and that makes the image look corrupt.

Fix that