plugins/obs-studio: Add behavior to toggle source action
This way we can always set a source to visible / unmute, or always set
it to hidden / mute, regardless of the current state. This can be
useful for example to create pre-stream actions where a certain
microphone or audio source starts disabled, or an end-of-stream button
that plays a song and always turns off microphone.