Introduce HTTP Request action
It lives under a new Network plugin.

There are 2 sources for the payload: a file in a well known
location; and in-memory text. The file is read every time the
action is activated, since we cannot monitor for file changes
due to FUSE limitations.

The in-memory text is editable through a small built-in dialog.
The Soup session is shared between all network action instances.